Refactored Styling in Vue components to live in CSS files
