+++ src/Shared/Clients/SpotifySearchClient.ts
