+++ src/Shared/NavBar/NavBar.css
@@ -1,0 +1,26 @@
+.navbar {
+    display: flex;
+    background-color: #C8D9E6;
+    align-items: center;
+    padding: 1rem;
+    font-size: 16px;
+
+    img {
+        height: 40px;
+        margin-right: 1rem;
+    }
+
+    ul {
+        list-style: none;
+        display: flex;
+        gap: 2rem;
+    }
+    
+    a {
+        color: #304674;
+        text-decoration: none;
+        padding: 0.5rem 1rem; /* Add padding inside the border */
+        border: 1px solid #304674; /* Border around each link */
+        border-radius: 4px; /* Optional rounded corners */
+    }
+}
