+++ src/components/EntryItem/EntryItem.css
@@ -1,0 +1,3 @@
+.entry-container {
+    
+}
