+++ src/components/Home/Home.css
@@ -1,0 +1,17 @@
+.home-container {
+    display: grid;
+    grid-template-columns: 2fr auto;
+    gap: 20px;
+
+    .left-column {
+        display: flex;
+        flex-direction: column;
+        gap: 20px;
+
+        .entries-container {
+            display: flex;
+            flex-direction: column;
+            gap: 20px;
+        }
+    }
+}
