+++ src/components/Logout/Logout.css
@@ -1,0 +1,3 @@
+.logout-container {
+    
+}
