+++ src/components/ProfileSection/ProfileSection.css
@@ -1,0 +1,16 @@
+.profile-section-container {
+    display: flex;
+    
+    .profile-picture {
+        width: 128px;
+        height: 128px;
+        margin-right: 20px;
+        border-radius: 50%; /* Makes the picture circular */
+    }
+
+    .profile-info p {
+        margin: 5px 0 ;
+        display: flex;
+        flex-direction: column;
+    }
+}
