+++ src/components/SpotifyUserLogin/SpotifyCallback.vue
@@ -36,5 +36,4 @@ onBeforeMount(() => {
         router.push({ name: 'Home' });
     }
 })
-
 </script>
