+++ src/components/SpotifyUserLogin/SpotifyUserLogin.css
@@ -1,0 +1,50 @@
+.container {
+    display: flex;
+    justify-content: center;
+    align-items: flex-start;
+    height: 100vh;
+    padding-top: 10vh; /* Adjusts the distance from the top */
+
+    /* Center-left alignment for title, description, and button */
+    .page-header {
+        max-width: 60%; /* Limit the width of content */
+        text-align: center;
+
+        /* Styling for the title */
+        .title {
+            font-size: 2.5rem;
+            font-weight: bold;
+            color: #304674;
+            margin-bottom: 0.5rem;
+        }
+
+        /* Styling for the description */
+        .description {
+            font-size: 1.2rem;
+            color: #304674; /* Gray text */
+            margin-bottom: 1.5rem;
+        }
+
+        /* Button styling */
+        .button-container {
+            display: flex;
+            justify-content: center;
+            margin-top: 1rem;
+
+            .login-button {
+                background-color: #6eb06e;
+                color: white;
+                padding: 12px 24px;
+                border: none;
+                border-radius: 25px;
+                font-size: 18px;
+                cursor: pointer;
+                transition: background-color 0.3s;
+            }
+        
+            .login-button:hover {
+                background-color: #5a8a5a; /* Darker blue on hover */
+            }
+        }
+    }
+}
